Update int to long

+++ LSMTree.Tests/DatabaseStressTest.cs
@@ -14,10 +14,10 @@ public class DatabaseStressTest
         var start = DateTime.Now;
         Console.WriteLine($"Inserting {limit:N0} records...");
 
-        for (int i = 0; i < limit; i++)
+        for (long i = 0; i < limit; i++)
         {
             db.Create(i, i);
-            int checkpoint = limit / 10;
+            long checkpoint = limit / 10;
             if (i % checkpoint == 0)
             {
                 Console.WriteLine($"Inserted: {i:N0}");
